requirements.txt
django==1.7
django-extensions
django-ordered-model
fabric
django-grappelli
icalendar
django-dotenv
django-getenv
dj-database-url
